Broke document viewer tests into discreet files
